セッション追跡のレビュー指摘に対応し、検索条件が効かない旨を明記する
- max_minutes を int に丸めてから範囲を確かめていたため、4294967306 などが
  桁あふれで 10 に化けて通っていた。long のまま SessionTrace で検証する
- 1 リクエストが 1000 行を超えて切られたとき、切った後ろの ID 行を起点に
  範囲を求め直し、境界付近の行が 2 つのリクエストに重複して出ていた。
  遡る途中で同じスレッドの直前のリクエストの範囲に入ったら止め、
  直前が上限で切られていれば同じリクエストに寄せる（追加の読み込みなし）
- 追跡中は検索条件欄（レベル・期間など）の絞り込みが効かないことを
  画面の制限事項と README に追記

+++ aplv-java/src/main/java/com/example/aplv/LogServer.java
@@ -778,9 +778,10 @@ public final class LogServer {
             sendErrorJson(ex, 400, "リクエストのはじまりとおわりを指定してください");
             return;
         }
-        int maxMinutes;
+        long maxMinutes;
         try {
-            maxMinutes = (int) parseLong(p.get("max_minutes"), SessionTrace.DEFAULT_MAX_MINUTES);
+            // int に丸めると桁あふれで範囲内の値に化けるため、long のまま範囲を確かめさせる
+            maxMinutes = parseLong(p.get("max_minutes"), SessionTrace.DEFAULT_MAX_MINUTES);
         } catch (NumberFormatException e) {
             sendErrorJson(ex, 400, "max_minutes は整数で指定してください");
             return;
